+++ internal/server/constant/enum.go
@@ -2,6 +2,21 @@
 
 package constant
 
+// ENUM(json, console)
+type ConfigLogEncoding string
+
+// ENUM(lowercase, lowercasecolor, capital, capitalcolor)
+type ConfigLogLevelEncoder string
+
+// ENUM(epoch, epochmillis, epochnanos, iso8601, rfc3339, rfc3339nano)
+type ConfigLogTimeEncoder string
+
+// ENUM(seconds, nanos, millis, string)
+type ConfigLogDurationEncoder string
+
+// ENUM(full, short)
+type ConfigLogCallerEncoder string
+
 // ENUM(basic_single, basic_credentials)
 type ConfigAuthMode string